Implement Result proto; refactor
