+++ src/PrimeFuncPack.Core/core_functionals_taggeds/Taggeds.Result/Result.T.Factory.cs
@@ -1,0 +1,8 @@
+﻿#nullable enable
+
+namespace System
+{
+    partial struct Result<TSuccess, TFailure>
+    {
+    }
+}
