+++ src/PrimeFuncPack.Core/core_functionals_taggeds/Taggeds.Result/Result.T.cs
@@ -2,7 +2,23 @@
 
 namespace System
 {
-    public readonly struct Result<TSuccess, TFailure>
+    public readonly partial struct Result<TSuccess, TFailure>
+        where TFailure : notnull, new()
     {
+        private readonly TaggedUnion<TSuccess, TFailure> _union;
+
+        private TaggedUnion<TSuccess, TFailure> Union => _union.IsInitialized switch
+        {
+            true => _union,
+            _ => TaggedUnion<TSuccess, TFailure>.CreateSecond(new TFailure())
+        };
+
+        private Result(in TSuccess success)
+            =>
+            _union = TaggedUnion<TSuccess, TFailure>.CreateFirst(success);
+
+        private Result(in TFailure failure)
+            =>
+            _union = TaggedUnion<TSuccess, TFailure>.CreateSecond(failure);
     }
 }
